+++ src/widgets/dropdown/lv_dropdown.c
@@ -19,6 +19,7 @@
 #include "../../misc/lv_anim.h"
 #include "../../misc/lv_math.h"
 #include "../../misc/lv_text_ap.h"
+#include "../../misc/lv_text_private.h"
 #include "../../stdlib/lv_string.h"
 #include <string.h>
