+++ src/widgets/roller/lv_roller.c
@@ -10,6 +10,7 @@
 #if LV_USE_ROLLER != 0
 
 #include "../../misc/lv_assert.h"
+#include "../../misc/lv_text_private.h"
 #include "../../draw/lv_draw.h"
 #include "../../core/lv_group.h"
 #include "../../indev/lv_indev.h"
