+++ src/widgets/span/lv_span.c
@@ -11,6 +11,7 @@
 #if LV_USE_SPAN != 0
 
 #include "../../misc/lv_assert.h"
+#include "../../misc/lv_text_private.h"
 #include "../../core/lv_global.h"
 
 /*********************
